feat: coupon entity 구현
refactor: 금액관련 자료형 전부 Long으로 변경

+++ src/main/java/coupon/domain/DiscountMoney.java
@@ -5,19 +5,19 @@ import lombok.Getter;
 @Getter
 public class DiscountMoney {
 
-    private static final int MINIMUM_DISCOUNT_AMOUNT_MONEY = 1000;
-    private static final int MAXIMUM_DISCOUNT_AMOUNT_MONEY = 10000;
-    private static final int DISCOUNT_AMOUNT_UNIT = 500;
+    private static final Long MINIMUM_DISCOUNT_AMOUNT_MONEY = 1000L;
+    private static final Long MAXIMUM_DISCOUNT_AMOUNT_MONEY = 10000L;
+    private static final Long DISCOUNT_AMOUNT_UNIT = 500L;
 
-    private final int discountAmount;
+    private final Long discountAmount;
 
-    public DiscountMoney(int discountAmount) {
+    public DiscountMoney(Long discountAmount) {
         validateDiscountAmountValue(discountAmount);
         validateDiscountAmountUnit(discountAmount);
         this.discountAmount = discountAmount;
     }
 
-    private void validateDiscountAmountValue(int discountAmount) {
+    private void validateDiscountAmountValue(Long discountAmount) {
         if (discountAmount < MINIMUM_DISCOUNT_AMOUNT_MONEY) {
             throw new IllegalArgumentException("할인 금액이 최소인 [%d] 금액보다 적습니다.".formatted(MINIMUM_DISCOUNT_AMOUNT_MONEY));
         }
@@ -27,7 +27,7 @@ public class DiscountMoney {
         }
     }
 
-    private void validateDiscountAmountUnit(int discountAmount) {
+    private void validateDiscountAmountUnit(Long discountAmount) {
         if (discountAmount % DISCOUNT_AMOUNT_UNIT != 0) {
             throw new IllegalArgumentException("할인 금액의 단위는 [%d] 로 설정해야 합니다.".formatted(DISCOUNT_AMOUNT_UNIT));
         }
