+++ src/main/java/coupon/domain/DiscountRate.java
@@ -6,22 +6,22 @@ import lombok.Getter;
 @Getter
 public class DiscountRate {
 
-    private static final int MINIMUM_DISCOUNT_RATE = 3;
-    private static final int MAXIMUM_DISCOUNT_RATE = 20;
+    private static final Long MINIMUM_DISCOUNT_RATE = 3L;
+    private static final Long MAXIMUM_DISCOUNT_RATE = 20L;
 
-    private final int discountRate;
+    private final Long discountRate;
 
     public DiscountRate(DiscountMoney discountMoney, OrderPrice orderPrice) {
-        int discountRate = (discountMoney.getDiscountAmount() * 100) / orderPrice.getPrice();
+        Long discountRate = (discountMoney.getDiscountAmount() * 100) / orderPrice.getPrice();
         validateDiscountRateRange(discountRate);
         this.discountRate = discountRate;
     }
 
-    public DiscountRate(int discountMoney, int orderPrice) {
+    public DiscountRate(Long discountMoney, Long orderPrice) {
         this(new DiscountMoney(discountMoney), new OrderPrice(orderPrice));
     }
 
-    private void validateDiscountRateRange(int discountRate) {
+    private void validateDiscountRateRange(Long discountRate) {
         if (discountRate < MINIMUM_DISCOUNT_RATE) {
             throw new IllegalArgumentException("할인율은 [%d] 이상이어야 합니다.".formatted(MINIMUM_DISCOUNT_RATE));
         }
